More query fixes

+++ fso-tables-worker/src/lib.rs
@@ -1154,12 +1154,12 @@ pub struct NewItem{
     pub item_text: String,
     pub documentation: String,
     pub major_version: String, 
-    pub parent_id: i64,
-    pub table_id: i64,
-    pub deprecation_id: i64,
-    pub restriction_id: i64,
+    pub parent_id: i32,
+    pub table_id: i32,
+    pub deprecation_id: i32,
+    pub restriction_id: i32,
     pub info_type: String,
-    pub table_index: i64,
+    pub table_index: i32,
     pub default_value: String,
 }
